base-hw: add very simple test-kernel
The test kernel is used for experiments with trustzone. It works on top of
the Versatile Express CoreTile Express A9x4 board only. The kernel initializes
the hardware and then idles in an endless loop. The only enabled interrupt is
the rx-irq of the uart3 serial line. Whenever that interrupt triggers a message
is printed on uart3.