Add 'add_region' method to vm-session interface.
The new add_region method enables a vmm to remove a certain ram-range from the
ram-allocator and adds it to the io-mem allocator. By that a vmm can attach the
appropriated physical memory for the trustzone VM to it's rm-session via an
io-mem dataspace.
Moreover, this commit fixes the prior senseless allocation of the vm_state
dataspace via an empty range_allocator. Instead it also uses the ram-allocator
for this.