Only use SRAM for the secure genode side, no DDR RAM.
The Trustzone Address Space Controller of the versatile express board only allows
to partition (and thereby secure) the SMC (static memory controller). The Trustzone
Protection Controller allows only full access or no access at all to the DMC
(including DDR RAM), so we can just decide whether to give the whole DDR RAM to
VM or nothing of it. By giving the whole DDR RAM to the VM, we have to take care,
that the whole secure world remains on the SMC side (e.g. SRAM of motherboard).