base & cortexa9: don't use FPU generally in Genode
If we enable the compiler to use FPU commands in every process it may does so
also inside the base-hw kernel. As the base-hw kernel must disable the FPU
by default to trap on user accesses (to enable lazy context switching), using
the FPU in the kernel is a bad idea. Hence, we disable FPU command generation
by default and leave it to the user to enable it for specific programs.

fix #1144