base: specs platform-panda_a and platform-panda_b
Support for the distinction of Panda Revison A and B at the level of the
platform specifier. This is done to enable base-hw later to consider the
correct board revision. However, at least base-foc does not need this
information, which is why the specifier platform_panda is still available and
references all aspects that are common to platform_panda_a and
platform_panda_b.

ref #1178