hw_panda: setup max MPU clock at initialization
This must be done inside the kernel as the speed of the Cortex-A9 timer
(scheduling timer) depends on the MPU clock. This implies also that
the MMIO of the clock manager 1 region A and the power-and-reset manager
must not be provided tothe user land as this would enable users to
influence the scheduling timing.

fix #1178