hw: provide Cpu_share::print_cpu_share_id
This enables us to set scheduling subjects into the global context of threads,
VMs and idle contexts when it comes to debugging. Although this isn't
necessary in the working system, it should be added permanently as it is a
pretty basic and often useful debugging feature which has no effect on the
binaries if not used.

fix #1300