libc VFS: FS for Linux frame buffer devices
The LXFB file system provides the dataspace of a Framebuffer session as
a Linux character-device file. The read and write operations are not
supported for that file but an mmap operation. In contrast to the
default mmap in Genodes Libc VFS that merely creates a copy of the
Dataspace content, the LXFB mmap returns the virtual address of the
original Dataspace after mapping the Dataspace locally.

Furthermore, an LXFB file supports two new IOCTLs to read the variable
as well as the fixed Screeninfo. The Screeninfos are set according to
the mode information of the underlying Framebuffer session.

Ref #1806