libports: Port avconv
The port of avconv is done via the already existing contrib code of
libav and the corresponding library ports. Additionally, as we want to
use avconv with the fbdev input format, we have to build libav with
avdevice in general. The commit adds a port of avdevice and adapts all
other users of libav to the change. The avdevice port supports only
fbdev by now. The commit adds the run/avconv test that records an
instance of eglgears for 30 seconds. The output is written to persistent
storage so that it can be accessed by the host system afterwards.

Ref #1806