hw: clean up scheduling state syscalls
This cleans up the syscalls that are mainly used to control the
scheduling state of a thread. The different use cases and requirements
were somehow mixed together in the previous interface. The new syscall
setup is:

1) pause_thread and resume_thread

They don't affect the state of the thread (IPC, signalling, etc.) but
merely decide wether the thread is allowed for scheduling or not, the
so-called pause state. The pause state is orthogonal to the thread state
and dominant. This feature is used as back end for the CPU session calls
"pause" and "resume" in applications that merely want to stop and
continue execution of a thread, like the GDB monitor. Both syscalls are
core-restricted and may target any thread.

2) stop_thread and restart_thread

The stop syscall can only be used on a thread in the non-blocking
("active") thread state. The thread then switches to the "stopped"
thread state in wich it explicitely waits for a restart. The restart
syscall can only be used on a thread in the "stopped" or the "active"
thread state. The thread then switches back to the "active" thread state
and the syscall returns whether the thread was in the stopped. Both
syscalls are not core-restricted. "Stop" affects the calling thread
while "restart" may target any thread in the same PD as the caller.
These syscalls are used when blocking for an event that is not known to
the kernel but when it is clear that the thread is active before and
after the blocking. Actually the syscalls are used when waiting for
locks and on thread exit.

3) cancel_thread_blocking

Does cleanly cancel any blocking thread state (including the "stopped"
thread state). The thread whose blocking was cancelled goes back to the
"active" thread state. It may receive a syscall return value that
reflects that its blocking was cancelled. This syscall doesn't affect
the pause state of the thread. It is core-restricted and may target any
thread.

4) yield_thread

Does its best that the calling thread is scheduled as few as possible in
the current scheduling superperiod without affecting the thread or pause
state. In the next superperiod, the thread is scheduled "normal" again.
The syscall takes another thread as argument but this has no effect. It
is not core-restricted.

Fixes #2104