avl_tree: provide No_match exception
Although the interface of the AVL tree does not provide a search method, the
main purpose of the AVL tree is performant searching. Each search algorithm,
build on top of the AVL tree needs a way of expressing that nothing was found.
This is the purpose of the No_match exception.

Ref #2139